Moved getting absolute URL into utils
